Increase enum size to allow for printing them out to console as shorts, not as chars

+++ include/OneLibrary/Constants.h
@@ -5,7 +5,7 @@
 
 namespace ol
 {
-	using InputInt = int32_t; // Type of packet. e.g. Mouse, Keyboard, Keep Alive.
-	using MouseInt = int32_t; // Mouse co-ordinate
-	using KeyboardInt = uint32_t; // Keyboard key flag (pressed, not pressed etc.) or keycode
+	using InputInt = int16_t; // Type of packet. e.g. Mouse, Keyboard, Keep Alive.
+	using MouseInt = int16_t; // Mouse co-ordinate
+	using KeyboardInt = uint16_t; // Keyboard key flag (pressed, not pressed etc.) or keycode
 }
